pyproject.toml
[project]
name = "hf-papers-zh-tw"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "pymupdf>=1.24",
    "anthropic>=0.40",
    "openai>=1.50",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "pillow>=10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
